apps/backend/requirements.txt
Django>=5.0,<6
djangorestframework>=3.14,<4
django-cors-headers>=4.3,<5
gunicorn>=21.0,<23
psycopg2-binary>=2.9,<3
